requirements.txt
torch>=2.0.0
transformers>=4.36.0
datasets>=2.14.0
peft>=0.7.0
trl>=0.7.4
accelerate>=0.24.0
wandb>=0.16.0
numpy>=1.21.0
sacrebleu>=2.3.0
rouge-score>=0.1.2
evaluate>=0.4.0
pyyaml>=6.0
tqdm>=4.64.0
scikit-learn>=1.2.0
gradio>=3.50.0
matplotlib>=3.5.0
seaborn>=0.11.0
pandas>=1.3.0